% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_canopies.R
\name{create_canopies}
\alias{create_canopies}
\title{Create canopies within a dataset}
\usage{
create_canopies(x, n_canopies, q)
}
\arguments{
\item{x}{a \code{matrix} or \code{data.frame} which can be coerced to a matrix
where each row represents a different record.}

\item{n_canopies}{An integer specifying how many canopies (ie- subsets) to create.}

\item{q}{An integer specifying the which norm to take the L-q distance of.}
}
\value{
A list of subset datasets
}
\description{
Compute high level splits of a dataset (ie- subsets or canopies)
based on L-q distance to the dataset centroid. Use of canopies allows for reducing
the time complexity of kNN by reducing it from 1 large \eqn{2^{O(n)}} problem
to several smaller problems.
}
\section{Details}{

The dataset is split into \code{n_canopies}. Canopies are created to be overlapping
with their neighbors. Specifically, let the calculated distance to the dataset centroid
be split into c ordered groups \eqn{c = 1, \ldots, n_canopies}. Define canopy 1 as all
observations of x with distance in \eqn{c_1 \cup c_2}, canopy n_canopies as all
observations of x with distance in \eqn{c_{n_canopies - 1} \cup c_{n_canopies}},
and all other canopies \eqn{j \in [2, n_canopies - 1]} as the observations of x
with distance in \eqn{c_{i-1} \cup c_i \cup c_{i+1}}.
}
\references{
McCallum, Andrew, Kamal Nigam, and Lyle H. Ungar.
"Efficient clustering of high-dimensional data sets with application to reference matching."
Proceedings of the sixth ACM SIGKDD international conference on Knowledge
discovery and data mining. ACM, 2000.
}

