% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R, R/int-impute_fn_knn.R
\name{impute_fn_knn}
\alias{impute_fn_knn}
\title{Imputation function for kNN}
\usage{
impute_fn_knn(values, distances, k, kern)

impute_fn_knn(values, distances, k, kern)
}
\arguments{
\item{values}{The values from which imputation will take place}

\item{distances}{The distances associated with each value}

\item{k}{The number of neighbors used to impute}

\item{kern}{The Gaussian kernal used for weighting}

\item{sigma}{The standard deviation (ie sigma parameter) of the Gaussian kernal used for weighting}

\item{values}{The values from which imputation will take place}

\item{distances}{The distances associated with each value}

\item{k}{The number of neighbors used to impute}
}
\description{
Function for KNN imputation for a single element.
Distances are weighted by a kernal function to produce a weighted
imputation.

Function for KNN imputation for a single element.
Distances are weighted by a kernal function to produce a weighted
imputation.
}
\seealso{
\code{\link[kernlab]{dots}}
}

