% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knn_impute.R
\name{kNN_impute}
\alias{kNN_impute}
\title{Imputation using weighted-kNN}
\usage{
kNN_impute(x, k, q = 2, verbose = TRUE, check.scale = TRUE,
  parallel = TRUE, leave_cores = 2)
}
\arguments{
\item{x}{a \code{matrix} or \code{data.frame} which can be coerced to a matrix
where each row represents a different record}

\item{k}{the number of neighbors to use for imputation}

\item{q}{An integer specifying the which norm to take the L-q distance of.}

\item{verbose}{if \code{TRUE} print status updates}

\item{check.scale}{Logical. If \code{TRUE} compute pairwise variance tests to see if
variables are on a common scale. Bonferroni correction applied.}

\item{parallel}{Logical. Do you wish to parallelize the code? Defaults to \code{TRUE}}

\item{leave_cores}{How many cores do you wish to leave open to other processing?}

\item{impute_fn}{The imputation function to run on the length k vector of values for
a missing feature.  Defaults to weighted mean-KNN; see Details.}
}
\description{
Imputation using weighted k-nearest neighbors.
For each record, identify missinng features.  For each missing feature
find the \eqn{k} nearest neighbors which have that feature.  Impute the missing
value using the \eqn{k} nearest neighbors having that feature. Weights are computed
using a Gaussian kernal bandwidth parameter using 'Silverman's rule of thumb'
as described by Silverman (1998)
}
\examples{
x = matrix(rnorm(100),10,10)
  x_missing = x > 1
  x[x_missing] = NA
  kNN_impute(x, 3)
}
\references{
Improved Methods for the Impution of Missing Data by Nearest Neighbors Methods
Tutz and Ramzan (2015)
}

