\name{meanImpute}
\alias{meanImpute}
\title{Mean Imputation
Fill missing values in a column with the mean of the column}
\usage{
  meanImpute(x)
}
\arguments{
  \item{x}{a data frame or matrix where each row represents
  a different record}
}
\description{
  Mean Imputation Fill missing values in a column with the
  mean of the column
}

