% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sort_indices}
\alias{sort_indices}
\title{Sort vector and return indexes}
\usage{
sort_indices(values)
}
\arguments{
\item{values}{The vector that should be sorted}
}
\description{
Sort the vector by the values
Return the indexes of the sorted vector according to original
}

