% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weighted_mean}
\alias{weighted_mean}
\title{Sort vector and return indexes}
\usage{
weighted_mean(x, w)
}
\arguments{
\item{values}{The vector that should be sorted}
}
\description{
Sort the vector by the values
Return the indexes of the sorted vector according to original
}

