% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_q}
\alias{dist_q}
\title{Calculate \eqn{L_q} distance of two vectors}
\usage{
dist_q(x, y, q)
}
\arguments{
\item{x}{A numeric vector. Missing values are allowed.}

\item{y}{A numeric vector. Missing values are allowed.}

\item{q}{An integer specifying the which norm to take the L-q distance of.}
}
\value{
a scalar
}
\description{
Calculate \eqn{L_q} distance of two vectors
}

