% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{impute_fn_knn}
\alias{impute_fn_knn}
\title{Imputation function for kNN}
\usage{
impute_fn_knn(values, distances, k, sigma)
}
\arguments{
\item{values}{The values from which imputation will take place}

\item{distances}{The distances associated with each value}

\item{k}{The number of neighbors used to impute}

\item{sigma}{The standard deviation (ie sigma parameter) of the Gaussian kernal used for weighting}
}
\description{
Function for KNN imputation for a single element.
Distances are weighted by a kernal function to produce a weighted
imputation.
}

